Move package core.job.sync.executor to core.execute
