+++ sharding-scaling-core/src/main/java/info/avalon566/shardingscaling/core/controller/SyncTaskController.java
@@ -21,10 +21,10 @@ import info.avalon566.shardingscaling.core.config.RdbmsConfiguration;
 import info.avalon566.shardingscaling.core.config.SyncConfiguration;
 import info.avalon566.shardingscaling.core.config.SyncType;
 import info.avalon566.shardingscaling.core.synctask.RealtimeDataSyncTask;
-import info.avalon566.shardingscaling.core.job.sync.executor.Event;
-import info.avalon566.shardingscaling.core.job.sync.executor.EventType;
-import info.avalon566.shardingscaling.core.job.sync.executor.SyncJobExecutor;
-import info.avalon566.shardingscaling.core.job.sync.executor.local.LocalSyncJobExecutor;
+import info.avalon566.shardingscaling.core.execute.executor.Event;
+import info.avalon566.shardingscaling.core.execute.executor.EventType;
+import info.avalon566.shardingscaling.core.execute.executor.SyncJobExecutor;
+import info.avalon566.shardingscaling.core.execute.executor.local.LocalSyncJobExecutor;
 import info.avalon566.shardingscaling.core.execute.reader.LogPosition;
 import info.avalon566.shardingscaling.core.execute.reader.ReaderFactory;
 import info.avalon566.shardingscaling.core.execute.util.DataSourceFactory;
