+++ sharding-scaling-core/src/main/java/info/avalon566/shardingscaling/core/execute/executor/Reporter.java
@@ -15,22 +15,23 @@
  * limitations under the License.
  */
 
-package info.avalon566.shardingscaling.core.job.sync.executor;
-
-import lombok.Data;
-import lombok.RequiredArgsConstructor;
+package info.avalon566.shardingscaling.core.execute.executor;
 
 /**
- * Event.
+ * Job Running reporter.
  * @author avalon566
  */
-@Data
-@RequiredArgsConstructor
-public class Event {
+public interface Reporter {
 
-    private final String taskId;
+    /**
+     * Report job running event.
+     * @param event job running event
+     */
+    void report(Event event);
 
-    private final EventType eventType;
-
-    private Object payload;
+    /**
+     * Get job running report.
+     * @return event
+     */
+    Event consumeEvent();
 }
