+++ sharding-scaling-core/src/main/java/info/avalon566/shardingscaling/core/execute/executor/SyncJobExecutor.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package info.avalon566.shardingscaling.core.job.sync.executor;
+package info.avalon566.shardingscaling.core.execute.executor;
 
 import info.avalon566.shardingscaling.core.config.SyncConfiguration;
 import info.avalon566.shardingscaling.core.controller.ReportCallback;
