+++ sharding-scaling-core/src/main/java/info/avalon566/shardingscaling/core/execute/executor/local/LocalReporter.java
@@ -15,10 +15,10 @@
  * limitations under the License.
  */
 
-package info.avalon566.shardingscaling.core.job.sync.executor.local;
+package info.avalon566.shardingscaling.core.execute.executor.local;
 
-import info.avalon566.shardingscaling.core.job.sync.executor.Event;
-import info.avalon566.shardingscaling.core.job.sync.executor.Reporter;
+import info.avalon566.shardingscaling.core.execute.executor.Event;
+import info.avalon566.shardingscaling.core.execute.executor.Reporter;
 
 import java.util.concurrent.ConcurrentLinkedQueue;
